Resolve different versions of the dcr alias
One salvaged from another set of scripts from days of yore

[d]ocker-[c]ompose [r]estart. Does a full clean restart of a running
docker-compose service by stopping it, removing it, and recreating it.

I rewrote it in one place while copying it from an old set of scripts in
another, so I'm just replacing it with the better of the two versions
here.